删除文章标签关联表，并修改文章和分类关系为一对一

+++ irh-service/irh-service-forum/src/main/java/top/imuster/life/provider/service/ArticleInfoService.java
@@ -144,7 +144,7 @@ public interface ArticleInfoService extends BaseService<ArticleInfo, Long> {
      * @param categoryId
      * @reture: top.imuster.common.base.wrapper.Message<top.imuster.life.api.pojo.ArticleInfo>
      **/
-    Message<List<ArticleInfo>> getBriefByCategoryId(Long categoryId, Long pageSize, Long currentPage) ;
+    Message<List<ArticleInfo>> getBriefByCategoryId(Long categoryId, Integer pageSize, Integer currentPage) ;
 
     /**
      * @Author hmr
